Moving to PyJWT as sanic-jwt is bad
It turns out that sanic-jwt is a bit depriciated and do not allow some functionalities, so I decided to write it myself.
