+++ gatewayApp/gateway/__main__.py
@@ -2,27 +2,21 @@
 Sanic app.
 """
 import argparse
-import hashlib
-import random
-import warnings
-from dataclasses import dataclass
 
 import sanic
 from sanic import response
 from sanic import request
 from sanic_ext import openapi
 from sanic_ext import validate
-from sanic_jwt import exceptions
-from sanic_jwt import initialize
-from sanic_jwt.decorators import protected
 
-from cryptography.hazmat.primitives.asymmetric import padding
-from cryptography.exceptions import InvalidKey
-from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives.serialization import load_pem_public_key
-from .node_connection import user_exists
-from .node_connection import add_public_key
-from .node_connection import get_public_key
+from .auth import protected
+from .auth import authenticate
+from .auth import register
+from .auth import generate_challenge
+
+from .valid_schemas import ValidUser
+from .valid_schemas import ValidChallengeRequest
+from .valid_schemas import ValidAuthRequest
 
 __version__ = '1.0.0'
 
@@ -37,86 +31,6 @@ def get_args():
 
     return argument_parser.parse_args()
 
-@dataclass
-class ValidUser:
-    user_id: str
-    public_key: str
-
-@dataclass
-class ValidChallengeRequest:
-    user_id: str
-
-def register(request_body):
-    """
-    Register a new user with user_id and his public key.
-    """
-    user_id = request_body.user_id
-    public_key_pem = request_body.public_key
-
-    if not user_id or not public_key_pem:
-        return response.json({"error": "user_id and public_key are required."}, status=400)
-
-    if user_exists(user_id):
-        return response.json({"error": "User ID already exists."}, status=400)
-
-    try:
-        public_key = load_pem_public_key(public_key_pem.encode())
-        add_public_key(user_id, public_key)
-    except (ValueError, InvalidKey):
-        return response.json({"error": "Invalid public key format"}, status=400)
-
-    return response.json({"message": "User registered successfully."}, status=201)
-
-
-def generate_challenge(request, request_body):
-    """
-    Generate a challenge for a given user_id.
-    Store the challenge in the app context.
-    Respond with the challenge and user_id.
-    """
-    user_id = request_body.user_id
-    if not user_id:
-        return response.json({"error": "Missing user_id."})
-
-    if not user_exists(user_id):
-        return response.json({"error": "User not found."})
-
-    challenge = hashlib.sha256(str(random.randint(0, 1_000_000)).encode()).digest()
-    request.app.ctx.challenges[user_id] = challenge
-
-    return response.json({"challenge": challenge.hex(), "user_id": user_id})
-
-def authenticate(request, *args, **kwargs):
-    """
-    Authenticate a user by verifying the signed challenge.
-    This function is called automatically by the sanic-jwt middleware on /auth endpoint.
-    When passed user can access endpoints protected by the @protected decorator
-    by providing a valid JWT token in the Authorization header.
-    """
-    user_id = request.json.get("user_id")
-    signed_challenge = bytes.fromhex(request.json.get("signed_challenge"))
-
-    challenge = request.app.ctx.challenges.pop(user_id, None)
-    if not challenge:
-        raise exceptions.AuthenticationFailed("Challenge not found for user.")
-
-    public_key = get_public_key(user_id)
-    if not public_key:
-        raise exceptions.AuthenticationFailed("User's public key not found.")
-
-    try:
-        public_key.verify(
-            signed_challenge,
-            challenge,
-            padding.PSS(
-                mgf=padding.MGF1(hashes.SHA256()),
-                salt_length=padding.PSS.MAX_LENGTH
-            ),
-            hashes.SHA256()
-        )
-    except Exception as exc:
-        raise exceptions.AuthenticationFailed("Challenge verification failed.") from exc
-    return {"user_id": user_id}
 
 def attach_endpoints(app):
     """Attach endpoints to the app."""
@@ -160,6 +74,20 @@ def attach_endpoints(app):
     def handle_challenge(request, body: ValidChallengeRequest):
         return generate_challenge(request, body)
 
+    @app.route('/auth', methods=["POST"])
+    @openapi.definition(
+        body=ValidAuthRequest,
+        summary="Authenticate a user and retrieve access and refresh tokens.",
+        response={
+            200: {"description": "Authentication successful, returns access and refresh tokens."},
+            400: {"description": "Error due to missing credentials."},
+            401: {"description": "Unauthorized, invalid username or password."}
+        }
+    )
+    @validate(json=ValidAuthRequest)
+    def handle_authentication(request, body: ValidAuthRequest):
+        return authenticate(request, body)
+
     @app.route('/protected')
     @openapi.definition(
         summary="Protected test endpoint.",
@@ -167,7 +95,7 @@ def attach_endpoints(app):
             200: {"description": "This is a protected endpoint."}
         }
     )
-    @protected()
+    @protected
     def protected_endpoint(request):
         return response.json({"message": "This is a protected endpoint."})
 
@@ -176,12 +104,7 @@ def create_app(arguments):
     app = sanic.Sanic("Gateway")
     app.ctx.args = arguments
     app.ctx.challenges = {}
-    app.config.SANIC_JWT_SECRET = "secret" #TODO: change this to a more secure secret
-
-    with warnings.catch_warnings(): # We have no event loop
-        warnings.simplefilter("ignore", category=DeprecationWarning)
-        initialize(app,
-            authenticate=authenticate)
+    app.config.SECRET = "secret" #TODO: change this to a more secure secret
 
     attach_endpoints(app)
     return app
