+++ gatewayApp/pytest.ini
@@ -1,5 +1,4 @@
 [pytest]
-; addopts = -p no:warnings
 filterwarnings =
     ignore::DeprecationWarning:websockets.connection
     ignore:pkg_resources is deprecated:DeprecationWarning
