+++ gatewayApp/requirements.txt
@@ -1,4 +1,4 @@
 sanic==24.6.0
-sanic-jwt==1.8.0
 sanic-ext==23.12.0
+PyJWT==2.9.0
 cryptography==43.0.3
